requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Graph Processing
networkx==3.2.1
python-igraph==0.11.3

# Deep Learning & GNN
torch==2.1.2
torch-geometric==2.4.0
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-cluster==1.6.3
dgl==1.1.3

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
streamlit==1.29.0
graphviz==0.20.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
loguru==0.7.2

# Data Download
kaggle==1.5.16

# Jupyter & Notebook Support
jupyter==1.0.0
ipykernel==6.27.1
ipywidgets==8.1.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==6.1.0
isort==5.13.2

# Model Monitoring
tensorboard==2.15.1
wandb==0.16.1

# Type Checking
mypy==1.7.1
